move ResetNamespace() function to utils_test.go
